AAE-1215
extract an interface for ProjectService
