+++ activiti-cloud-services-org/activiti-cloud-services-org-service/src/main/java/org/activiti/cloud/services/organization/service/api/ProjectService.java
@@ -1,0 +1,34 @@
+package org.activiti.cloud.services.organization.service.api;
+
+import org.activiti.cloud.organization.api.Project;
+import org.activiti.cloud.services.common.file.FileContent;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
+import org.springframework.lang.Nullable;
+import org.springframework.web.multipart.MultipartFile;
+
+import java.io.IOException;
+import java.util.Optional;
+
+/**
+ * Business logic related to {@link Project} entities
+ */
+public interface ProjectService {
+    Page<Project> getProjects(Pageable pageable,
+                              String name);
+
+    Project createProject(Project project);
+
+    Project updateProject(Project projectToUpdate,
+                          Project newProject);
+
+    void deleteProject(Project project);
+
+    Optional<Project> findProjectById(String projectId);
+
+    FileContent exportProject(Project project) throws IOException;
+
+    Project importProject(MultipartFile file, @Nullable String name) throws IOException;
+
+    void validateProject(Project project);
+}
